[Refactoring] Package Name 변경
option.printOption -> option.print
option.compareOption -> option.compare
