+++ src/main/java/option/print/CountPrintOption.java
@@ -1,4 +1,4 @@
-package option.printOption;
+package option.print;
 
 import employee.Employee;
