+++ src/test/java/command/AddCommandTest.java
@@ -1,12 +1,11 @@
 package command;
 
-import command.AddCommand;
 import database.Database;
 import employee.Employee;
 import exceptions.ImproperlyConfigured;
 import option.Option;
-import option.compareOption.DefaultCompareOption;
-import option.printOption.CountPrintOption;
+import option.compare.DefaultCompareOption;
+import option.print.CountPrintOption;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
