+++ src/test/java/database/SelectTest.java
@@ -2,7 +2,7 @@ package database;
 
 import employee.Employee;
 import exceptions.ImproperlyConfigured;
-import option.compareOption.*;
+import option.compare.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
